refactor: inline tests
